[arch/linux] separate file for get_rt_prio and some cleanup
